Refactor flow stores and centralize fallback handling
