+++ webapp/src/services/fallbackRegistry.js
@@ -1,0 +1,1 @@
+export * from './fallbackRegistry.ts';
